backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
pymongo==4.6.1
djongo==1.3.6
mongo-driver
django-cors-headers==4.3.0
drf-yasg==1.21.7
python-decouple==3.8
bcrypt==4.0.1
sqlparse==0.4.4
